complete TasbihCounter
