+++ Tasbih Counter/Tasbih Counter/AppDelegate.swift
@@ -1,6 +1,6 @@
 //
 //  AppDelegate.swift
-//  TasbihCounter
+//  Tasbih Counter
 //
 //  Created by Ameera BA on 11/10/2021.
 //
